Add missing migration init
