+++ core/forms.py
@@ -1,7 +1,5 @@
 from django import forms
 from django.core.validators import MaxLengthValidator, URLValidator
-from django.utils.text import slugify
-
 import os
 from urllib.parse import urlparse
 
@@ -10,7 +8,7 @@ import mistune
 import requests
 from PIL import Image
 
-from .models import Community
+from communities.models import Community
 from .utils.link_safety import check_url_safety
 
 
@@ -159,26 +157,3 @@ class PostForm(forms.Form):
         return cleaned
 
 
-class CommunityCreateForm(forms.ModelForm):
-    class Meta:
-        model = Community
-        fields = ["slug", "name", "title", "description"]
-
-    def clean_slug(self):
-        slug = slugify(self.cleaned_data.get("slug", ""))
-        MaxLengthValidator(191)(slug)
-        if not slug:
-            raise forms.ValidationError("Slug is required.")
-        if Community.objects.filter(slug=slug).exists():
-            raise forms.ValidationError("Community with this slug already exists.")
-        return slug
-
-    def clean(self):
-        cleaned_data = super().clean()
-        name = cleaned_data.get("name")
-        if name is not None:
-            name = name.strip()
-            if Community.objects.filter(name__iexact=name).exists():
-                self.add_error("name", "Community with this name already exists.")
-            cleaned_data["name"] = name
-        return cleaned_data
