+++ core/tests/test_astro.py
@@ -10,7 +10,8 @@ from django.urls import reverse
 from django.utils import timezone
 from freezegun import freeze_time
 
-from core.models import Community, EngagementEvent, Post, CommunityBaseline
+from communities.models import Community, CommunityBaseline
+from core.models import EngagementEvent, Post
 from votes.services import cast_vote_post_once, AlreadyVoted
 from core.services.astro import compute_post_signals
