+++ core/tests/test_comment_edit.py
@@ -6,7 +6,8 @@ from django.test import TestCase
 from django.urls import reverse
 from django.utils import timezone
 
-from core.models import Community, Post
+from communities.models import Community
+from core.models import Post
 from comments.models import Comment
