+++ core/tests/test_deleted_post_visibility.py
@@ -2,7 +2,8 @@ from django.contrib.auth import get_user_model
 from django.test import TestCase
 from django.urls import reverse
 
-from core.models import Community, Post
+from communities.models import Community
+from core.models import Post
 
 
 class DeletedPostVisibilityTests(TestCase):
