+++ core/tests/test_feed_time_filters.py
@@ -4,7 +4,8 @@ from django.test import TestCase
 from django.urls import reverse
 from django.utils import timezone
 
-from core.models import Community, Post
+from communities.models import Community
+from core.models import Post
 
 
 class FeedRangeMixin:
