+++ core/tests/test_post_excerpt.py
@@ -1,7 +1,8 @@
 from django.contrib.auth import get_user_model
 from django.test import TestCase
 
-from core.models import Community, Post
+from communities.models import Community
+from core.models import Post
 
 
 class PostExcerptTests(TestCase):
