+++ core/tests/test_post_image_link.py
@@ -2,7 +2,8 @@ from django.contrib.auth import get_user_model
 from django.core.exceptions import ValidationError
 from django.test import TestCase
 
-from core.models import Community, Post, PostImageLink
+from communities.models import Community
+from core.models import Post, PostImageLink
 
 
 class PostImageLinkTests(TestCase):
