+++ core/tests/test_transparency_posts.py
@@ -4,7 +4,8 @@ from django.contrib.auth import get_user_model
 from django.test import TestCase, override_settings
 from django.urls import reverse
 
-from core.models import Community, Post
+from communities.models import Community
+from core.models import Post
 from votes.services import cast_vote_post_once
