+++ tests/test_login_redirect.py
@@ -3,7 +3,8 @@ from django.contrib.auth import get_user_model
 from django.urls import reverse
 from urllib.parse import quote
 
-from core.models import Community, Post
+from communities.models import Community
+from core.models import Post
 
 
 @pytest.mark.django_db
