+++ tests/test_votes.py
@@ -1,7 +1,8 @@
 import pytest
 from django.urls import reverse
 
-from core.models import Community, Post
+from communities.models import Community
+from core.models import Post
 
 
 @pytest.fixture
