+++ votes/tests/test_vote_engagement.py
@@ -1,7 +1,8 @@
 from django.contrib.auth import get_user_model
 from django.test import TestCase
 
-from core.models import Community, EngagementEvent, Post
+from communities.models import Community
+from core.models import EngagementEvent, Post
 from votes.services import cast_vote_post_once
